Update patch set 68
Patch Set 68:

> Which binary digest?

The byte array that should be returned from finalize() - it's not a state that needs to live for the entire lifetime of the digest object.

> Dees_Troy said it was not necessary to bump the theme version.

He was mislead by your misleading comment. It is needed because you modified the variable names.

> The refactor makes the digest class more simple, and puts more processing in the callee.

That's exactly the opposite of good design. I suggested a separate digest driver class since you pushed the first version of this to Gerrit, but you ignored me.

Patch-set: 68