Update en.xml
I discovered several inconsistencies in the wording of "SD Card". All text strings state "SD Card" - with a space between "SD" & "Card" and a capital "C", except 3. I corrected these 3 strings to "SD Card" (with a space and capital "C") to be consistent with all other strings. I would be happy to see them corrected in the upcoming patch/release.