Update patch set 2
Patch Set 2: Code-Review+1

Patch works, and feels more complete than the previous iteration. Only issue with this one (and it's livable) is that in the log, when the early time fix takes place, the time zone won't match the user's setting as settings haven't been loaded yet (and may not be available if /data is encrypted). It may be worth getting this other patch working too: https://gerrit.omnirom.org/c/24386/

That would allow settings to be loaded early as well. At any rate, this patch fixes the current issue on my Oreo device that needs the Qualcomm fix, so I say ship it, and build on it later :).

Patch-set: 2
Reviewer: Gerrit User 1000092 <1000092@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=+1