Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)

The normal TWRP settings file (irrelevant of where it's stored), will always be read after this, and will therefore be used as usual.

Basically this is just a backup file (which get's saved at the same time as the normal settings file), and it's primary function is to allow for 'early' reading of the settings, and use a few variables for better user experience.
As well as account for qcom time fix where it's only stored /persist/time/ats_* and no longer in /data.

It should in no way affect normal TWRP operation or settings.

Patch-set: 3