Update patch set 5
Patch Set 5:

sorry for being late on this topic and good to see it get merged. but can you confirm this: for old device system don't support /persist/time (but do have /persist) TWRP will use /data/time/ as top priority? having a setting value to make it work in early boot is nice but it shouldn't override for system that uses only /data/time/ because the old system may only (and always) update /data/time/ which maybe different from twrp's backup of this value

Patch-set: 5