Update patch set 3
Patch Set 3: Code-Review+1

Updated based on _that's feedback (thanks!) and added the new function to the PartitionManager class so I can call it from the vold_decrypt code. Seems be be working the same as before and the code is much cleaner. Also updated the style of the code to be more consistent with the existing stuff.

Patch-set: 3
Label: Code-Review=+1