Adding Edl button in reboot menu
Edl mode is available only on qualcomm devices and it used to fully 
restore the bricked devices.
I make the button to be available only if TW_HAS_EDL_MODE := true like download button.
I setted the edl button in the same position of the download button, because the device that have the download mode doesn't have the edl mode and viceversa.
I also added the translation for all language.


Change-Id: I69d1058b8c050b722395c7b3da6dce1ccfbe7a91