Update patch set 3
Patch Set 3:

There was a conversation about this here: https://gerrit.omnirom.org/#/c/android_bootable_recovery/+/31021/

If there's a way to handle this automatically, that would be preferred. We need to see if automatic detection causes issues on any legacy devices, or if there are devices with this new type of vibrator and the legacy type.

Patch-set: 3
Label: Verified=0