Update patch set 1
Patch Set 1:

Additionally, "skip twrp built-in decryption if using system vold to save time" needs to be undone. vold decrypt is a fallback for TWRP crypto, not a replacement. For some devices, depending on the ROM/Android version installed, one or the other may be needed. They both always need to be able to co-exist. Please update the patch to remove this change.

Patch-set: 1