Update patch set 2
Patch Set 2:

> Disabled soong build for vdc_pie, building with make and only if
 > 9.0 android tree is detected. Also removed twrp built-in decryption
 > skip as suggested.
With the latest patch set, no vdc_pie binary is produced in recovery/root/sbin at all. And how will the binary be produced if an older Android tree is used for the build? Just because I build in the 8.1 tree doesn't mean I won't need to decrypt a Pie ROM.

Patch-set: 2