WIP: Use magiskboot to repack the boot partition
Use magiskboot and provide GUI options to allow users to repack
their existing boot image to install TWRP (or kernels) so we can
stop having to provide installation zips for AB devices. There is
also an option to try to fix a recovery bootloop if the kernel
has been patched to always boot the ramdisk for root, etc.

Change-Id: I74196cc6f095a7576d61886dc96cbc18deba9b04