WIP: Use magiskboot to repack the boot partition
Use magiskboot and provide GUI options to allow users to repack
their existing boot image to install TWRP (or kernels) so we can
stop having to provide installation zips for AB devices.

Could also offer options to remove forceencrypt and dm-verity on
other devices, but not sure if we should enable those features
through build flags or with fstab flags.

Change-Id: I74196cc6f095a7576d61886dc96cbc18deba9b04