Update patch set 6
Patch Set 6: Code-Review-1

Here's a problem. vdc_pie is only building if the android-9.0 tree is detected, but what about devices that build in an older tree but need to decrypt a Pie ROM? There needs to be a way to allow it to be built no matter what tree is being used. My device is failing to decrypt Pie because I built in the 8.1 tree.

Patch-set: 6