Update patch set 6
Patch Set 6:

> (1 comment)
Even if changing that to 27 gets it to build on Oreo, what about recovery builds in older trees? The idea behind vold_decrypt was to use the system binaries so that it wouldn't matter which tree the recovery was being built in - the binaries would always be the correct ones for the ROM. Do we really need to build vdc_pie at all, or can we just use system vdc and call the newer commands based on the detected sdk version? That would probably be the best solution here.

Patch-set: 6