Update patch set 10
Patch Set 10:

Not sure how close this patch is to complete, but after extensive testing, the decryption routine appears to be working on Pie for me (haven't tested Oreo). There are a couple of odd behaviors not completely related to this patch, though:

- When building with busybox, decrypt times out every time. Decrypt works fine when building with toybox.

- I had to comment out the "ldconfigtxt" service from the init.recovery.service.TV file, as it seems to delay boot and again causes decryption to timeout.

When building with toybox and commenting out that service, decryption works reliably on the devices I'm building for. If either of those two things is done (building with busybox OR leaving that service enabled), decryption becomes unreliable and I end up having to use vold decrypt as a fallback.

Patch-set: 10