Update patch set 3
Patch Set 3:

This is causing some delays/failures in decryption on FDE devices. I recommend a revert and let device maintainers handle this directly, without the associated service. Another alternative could be to disable the service and run manually "on fs" and add a "wait /sbin" to ensure the file system is available. I'm not sure if the latter will fix the issue, but currently I'm not sure when this service runs, since it's not associated with a class or a trigger that I see.

Patch-set: 3