Update patch set 4
Patch Set 4:

The last major theme update was 2 years ago, but minor tweaks are made periodically including the adjustments that I'm currently making to the advanced menu (https://gerrit.omnirom.org/#/c/android_bootable_recovery/+/33236/). I think we're going to have to agree to disagree on this theme as I don't see a need for a separate notch theme. The build flag provides an acceptable workaround though I personally never even thought that using the build flag was a necessity. IMHO, notches are already on their way out. 2018 was the year of the notch and 2019 is starting to look like the year of the punch hole. In another year or 2, notches will probably be a thing of the past for most devices. Another point of concern that I have for this theme is the fact that different notches are different heights and sizes and I don't see anywhere that this theme addresses that issue (consider how tall the notch is on the Pixel 3 XL). The build flag allows the device maintainer to adjust the height for each device.

Patch-set: 4
Reviewer: Gerrit User 1000004 <1000004@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=0