Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> I am abandoning this change. It is getting too big and device specific...
> You can take over.
Who can take over? I'm not a coder; I'm just a tester. The items I reported aren't device specific at all - they're just related to the order of operations when decrypt fails. All that needs to change is for the services to be stopped after a failed decrypt just like they are for a successful one. I'm not sure I understand why you would abandon a change due to feedback. This is something that anyone that uses/needs vold decrypt will need. Please reconsider giving up on it.

Patch-set: 10