Update patch set 10
Patch Set 10:

One last note - I tried cancelling the prompt after an incorrect password on my U11 running Pie, and I do see the services stopping as expected in the log. When I go to manually decrypt, however, TWRP crypto fails, and when I look in the log, it doesn't appear that all of the TWRP services restarted after the failed decrypt. I think that may be more related to the order/triggers for the services in my rc file, though. I'm going to look into moving those around a bit and see if it makes any difference for the failing TWRP crypto after cancel. Either way, I would agree that it's not the fault of or related to this patch.

Patch-set: 10