Update patch set 10
Patch Set 10:

(1 comment)

> Patch Set 10:
> 
> (1 comment)
> 
> Scenario #3; That sounds about right. I don't use TWRP decryption at all(disabled) so I don't have those problems. TWRP and vold decryptions gets on each others way as you can see. I can move unmount and stop-services call somewhere else to make sure it always executes once you reach twrp main screen(not just with "Cancel" press)
> Scenario #4: That's gonna be tricky... As I said earlier if I let vold_decrypt to stop its services, unmount partitions and restart twrp services after failed decryption with wrong password then following retries with good password never succeed, at least on my device tested with both Oreo and Pie roms. That's why I let those things running till decryption success or Cancel pressed. I can't think of anything that can make this scenario working. But you can eliminate it easy by using only one method of decryption - vold or twrp. If twrp works don't use vold. Use vold only if twrp can't decrypt, twrp won't get on vold's way then so should be good, no scenario 4 should happen...

Unfortunately on many HTC devices, vold decrypt is required on stock, but TWRP crypto is needed on AOSP-based ROMs. There are also cases where older Android versions for the same device need TWRP crypto but newer versions need vold decrypt. So the two need to co-exist, especially since the original intent of vold decrypt was to be a fallback and not a replacement. The change you had me make to the code though seems to have resolved this issue.

> 
> > So my U12+ tester has confirmed that decryption still fails after
> > inputting incorrect password, then cancelling and inputting correct
> > password. However, I don't believe that failure is a result of this
> > patch, so Scenario #4 can be disregarded. The one thing I did
> > notice in his recovery log is that is doesn't seem that vdc_pie is
> > attempting to decrypt at all (at least I don't see any log output
> > from it).
> Yeah, I noticed that too, tried to debug it briefly but looks like with wrong password vdc_pie doesn't get any message from vold in the 30 sec timeout so it doesn't return anything to vold_decrypt to log. I'll see if I can dig more...

It's not that important if vdc_pie isn't producing logging, as long as it's working. But when it's not, the lack of logging makes it difficult to know why it failed. The even better news with all of this is that if you just enter the correct password in the first place, everything works, and at worst, if you can't get it working after an incorrect password, rebooting recovery and inputting the correct one should still work.

Patch-set: 10