Update patch set 11
Patch Set 11:

> Patch Set 11: Code-Review+1
> 
> I think I got right this time... Reverted last change, added 1 second delay before next vold service start and everything seems to work great on every decryption try/retry. Phew...

The additional delay that was added seems to be preventing the keymaster-3-0 service from starting on at least one of the devices I'm building for. What's weird is that the log says the service is running, but then later it says it's not and resumes the TWRP version of the service. I know that the keymaster service can be very touchy when it comes to timing, so I'm wondering if the sleep that was added is too large?

Here's an excerpt from the log so you can see what I mean: http://tinyw.in/6Cqu

Patch-set: 11