Update patch set 11
Patch Set 11:

> Patch Set 11: -Code-Review
> 
> qseecomd should start first to have time to initialize all listeners ready. I actually was thinking to ensure that in the code, I still may do it.
> My order is qseecomd, hwservicemanager, keymaster-3-0, servicemanager and have no issues.
> But you are right in general - I haven't played with the delay alto it worked fine with 5 seconds on my device testing it the first time. Recovery log doesn't help here, get logcat and maybe dmesg rifht after decryption fail to see what's the keymaster problem. On my device without the delay kernel hw binder wasn't able to bind kaymaster service the second time hence IKaymasterDefault HAL did't start. 
> Maybe I should do delay per device configurable, something like TW_SYSTEM_VOLD_SERVICES_DELAY := 500 in ms or something!?

Making it configurable per device would be a nice idea, with some value as the default if it's not set. And if you could add some code to ensure that sys.listeners.registered=true before trying to start the other services, then there shouldn't need to be a sleep in there at all, right? In my rc file for the TWRP services, I have the keymaster service starting on hwservicemanager.ready=true as well. So maybe incorporating those into the code would make it more reliable.

Patch-set: 11