Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> I tried before wait for prop sys.listeners.registered=true after qseecomd start but it didn't work.
> Let me see what else I can do...

Is it possible for the vold decrypt code to be updated to unset those specific props when it stops the TWRP and/or vold decrypt services? That's the actual reason that TWRP crypto isn't working after a bad password attempt. vold decrypt restarts the TWRP services, but in fact keymaster isn't ready because qseecomd and hwservicemaster aren't really ready. But since the sys.listeners.registered=true and hwservicemanager.ready=true props stay set even after the services are stopped, there's no way for keymaster to know to wait. I'm thinking about this now, and maybe we can set those props in the service themselves, like this?

https://github.com/omnirom/android_bootable_recovery/blob/android-9.0/crypto/vold_decrypt/init.recovery.vold_decrypt.hwservicemanager.rc#L15

I'm not sure if there's an "onstop" option for init services, though.

Patch-set: 11
Label: Code-Review=0