Update patch set 13
Patch Set 13:

> Patch Set 13:
> 
> These goes in your init.qcom.rc or whatever with the twrp decrypt services

Like I already said, I already have that stuff covered in my rc file. That's not the problem. The problem is the code I linked that restarts the services. Actually, if it didn't restart the keymaster service, the rc file would take care of it, but since it does, the TWRP service gets stuck in "restarting" status. So the rc file can't start the service because it's already "restarting" and never gets to running status, because when it was started, the proper services/props weren't ready. I must not be explaining this well. Maybe all we need to do is blacklist the keymaster-3-0 service from being restarted by vold decrypt when it's finished? I think that would solve the problem.

Patch-set: 13
Label: Code-Review=0